Added check for amp===0 for weighted pool. Small getBestPathIds fix. Multicall to use lower case address. Started trade script.
